etl/requirements.txt
confluent-kafka>=2.3.0
pydantic>=2.0.0
httpx>=0.25.0
python-dotenv>=1.0.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp-proto-grpc>=1.22.0
